use integer instead of float at FunctionResolver#queryVersion

+++ libvpvl2/src/engine/nvfx/EffectContext.cc
@@ -100,7 +100,7 @@ struct FunctionResolverProxy : nvFX::FunctionResolver {
     void *resolve(const char *name) const {
         return m_resolver->resolveSymbol(name);
     }
-    float queryVersion() const {
+    int queryVersion() const {
         return m_resolver->query(IApplicationContext::FunctionResolver::kQueryVersion);
     }
